[fix] 회원가입 redux변경 및 오류 수정
