+++ src/redux/modules/authSlice.js
@@ -5,17 +5,20 @@ import * as authApi from "@api/authApi";
 const initialState = {
     isLogin: false,
     userInfo: null,
-    errorMessage: "",
+    authMessage: "",
+    fetchRegisterState: "",
+    fetchLoginState: "",
+    isRegistCompleted: false,
 };
 
 export const register = createAsyncThunk(
     "auth/register",
     async (formData, { rejectWithValue }) => {
         try {
-            const response = authApi.signup(formData);
-            console.log(response);
+            const response = await authApi.signup(formData);
+
+            return response.data;
         } catch (error) {
-            console.log(error.response.data);
             return rejectWithValue(error.response.data);
         }
     }
@@ -24,11 +27,12 @@ export const login = createAsyncThunk(
     "auth/login",
     async (formData, { rejectWithValue }) => {
         try {
-            const response = authApi.signin(formData);
-            const { accessToken, refreshToken } = response.data;
-            TokenService.setUserData(accessToken, refreshToken);
+            const response = await authApi.signin(formData);
+            const userId = response.data.data.userId;
+            const userInfo = await authApi.getUserData(userId);
+            return { response: response.data, userInfo: userInfo.data.data };
         } catch (error) {
-            console.log(error.response.data);
+            console.log(error);
             return rejectWithValue(error.response.data);
         }
     }
@@ -38,25 +42,49 @@ const authSlice = createSlice({
     name: "auth",
     initialState,
     reducers: {
-        // isLoggedin(state, action) {
-        //     state.isLogin = action.payload;
-        // },
-        // logOut(state, action) {
-        //     TokenService.removeUser();
-        //     state.isLogin = false;
-        //     state.userInfo = null;
-        // },
-        // // 유저 닉네임 변경
-        // setUserNickname(state, action) {
-        //     state.userInfo.nickname = action.payload;
-        // },
-        // // 유저 프로필 이미지 변경
-        // setUserProfileImage(state, action) {
-        //     state.userInfo.profileImageUrl = action.payload;
-        // },
+        logOut(state, action) {
+            TokenService.removeUserData();
+            state.isLogin = false;
+            state.userInfo = null;
+        },
+        // 유저 닉네임 변경
+        setUserNickname(state, action) {
+            state.userInfo.nickname = action.payload;
+        },
+        // 유저 프로필 이미지 변경
+        setUserProfileImage(state, action) {
+            state.userInfo.profileImageUrl = action.payload;
+        },
     },
-    extraReducers: (builder) => {},
+    extraReducers: (builder) => {
+        builder.addCase(register.pending, (state) => {
+            state.fetchRegisterState = "LOADING";
+        });
+        builder.addCase(register.fulfilled, (state, action) => {
+            state.fetchRegisterState = "SUCCESS";
+            state.authMessage = action.payload.message;
+            state.isRegistCompleted = true;
+        });
+        builder.addCase(register.rejected, (state, action) => {
+            state.fetchRegisterState = "ERROR";
+            state.authMessage = action.payload.message;
+        });
+        builder.addCase(login.pending, (state) => {
+            state.fetchLoginState = "LOADING";
+        });
+        builder.addCase(login.fulfilled, (state, action) => {
+            state.isLogin = true;
+            state.fetchLoginState = "SUCCESS";
+            state.authMessage = action.payload.response.message;
+            state.userInfo = action.payload.userInfo;
+        });
+        builder.addCase(login.rejected, (state, action) => {
+            state.fetchLoginState = "ERROR";
+            state.authMessage = action.payload.message;
+        });
+    },
 });
 
-export const { logOut } = authSlice.actions;
+export const { logOut, setUserNickname, setUserProfileImage } =
+    authSlice.actions;
 export default authSlice.reducer;
