--- src/redux/modules/msgSlice.js
@@ -1,19 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-
-const initialState = {
-    message: {},
-};
-
-const msgSlice = createSlice({
-    name: "error",
-    initialState,
-    reducers: {
-        showModal(state, action) {
-            state.message = action.payload;
-        },
-    },
-});
-
-export const { showModal } = msgSlice.actions;
-
-export default msgSlice.reducer;
